+++ angular/services/toast.js
@@ -1,0 +1,40 @@
+(function(){
+	"use strict";
+
+	angular.module("app.services").factory('ToastService', function($mdToast){
+
+		var delay = 6000,
+			position = 'top right',
+			action = 'OK';
+
+		return {
+			show: function(content){
+				if (!content){
+					return false;
+				}
+
+				return $mdToast.show(
+					$mdToast.simple()
+						.content(content)
+						.position(position)
+						.action(action)
+						.hideDelay(delay)
+				);
+			},
+			error: function(content){
+				if (!content){
+					return false;
+				}
+
+				return $mdToast.show(
+					$mdToast.simple()
+						.content(content)
+						.position(position)
+						.theme('warn')
+						.action(action)
+						.hideDelay(delay)
+				);
+			}
+		};
+	});
+})();